refactor(web): relocate components out of sidebar dir, rename it to nested-safes as per content
