+++ apps/web/src/components/nested-safes/NestedSafesList/SimilarityGroupContainer.tsx
