+++ apps/web/src/components/nested-safes/NestedSafesPopover/index.tsx
@@ -17,14 +17,14 @@ import SettingsIcon from '@/public/images/sidebar/settings.svg'
 import { ModalDialogTitle } from '@/components/common/ModalDialog'
 import { CreateNestedSafeFlow } from '@/components/tx-flow/flows'
 import { TxModalContext } from '@/components/tx-flow'
-import { NestedSafesList } from '@/components/sidebar/NestedSafesList'
-import { NestedSafeInfo } from '@/components/sidebar/NestedSafeInfo'
-import { NestedSafeIntro } from '@/components/sidebar/NestedSafeIntro'
+import { NestedSafesList } from '@/components/nested-safes/NestedSafesList'
+import { NestedSafeInfo } from '@/components/nested-safes/NestedSafeInfo'
+import { NestedSafeIntro } from '@/components/nested-safes/NestedSafeIntro'
 import Track from '@/components/common/Track'
 import { NESTED_SAFE_EVENTS } from '@/services/analytics/events/nested-safes'
 import CheckWallet from '@/components/common/CheckWallet'
-import { useManageNestedSafes } from '@/components/sidebar/NestedSafesList/useManageNestedSafes'
-import { SimilarityConfirmDialog } from '@/components/sidebar/NestedSafesList/SimilarityConfirmDialog'
+import { useManageNestedSafes } from '@/components/nested-safes/NestedSafesList/useManageNestedSafes'
+import { SimilarityConfirmDialog } from '@/components/nested-safes/NestedSafesList/SimilarityConfirmDialog'
 import type { NestedSafeWithStatus } from '@/hooks/useNestedSafesVisibility'
 
 function PopoverHeaderAction({
